Bump versions on plugins and recreate infrastructure
Before the break I had removed a bunch of dangling infrastructure
(including deleting the default VPC for us-west-2). These changes
reflect the differences when I came back. The aws-vpc module was having
trouble because of the deleted VPC so I ended up deleting the terraform
state file from S3. This wasn't quite enough to allow `terragrunt init`
to work because there was a dangling lock in dynamodb. After deleting
that `terragrunt init/plan` seems to be happy and we are back on the
road.

Signed-off-by: Ian Allison <iana@pims.math.ca>