Attempting to catch breaking changes in update
This is the first time we've deployed with Helm3 and things are likely a
bit wonky. For one thing we were breaking the io.k8s.api.core.v1.PodSpec
by giving a name in our deployment spec. The README.md is being updated
with some of the differences and some requirements are being updated.

Signed-off-by: Ian Allison <iana@pims.math.ca>