Set ExternalTrafficPolicy on shib service
It seems to be necessary to have Local as the ExternalTrafficPolicy.
Ideally the value Cluster provides better load balancing at the expense
of introducing an extra hop in the network, but this doesn't seem to
work for us. Using the Cluster policy curl returns empty responses. This
is probably a broken configuration on our side somewhere but because
we're not load balancing shibboleth yet it doesn't really matter, we
only have one pod available.

Signed-off-by: Ian Allison <iana@pims.math.ca>