Move to syzygy.farm domain
The domain is still hard coded in the image in a couple of places,
specifically, we do a redirect to the domain to go from http->https and
we have a couple of references to the domain in our html. The html
should be an easy fix, but the redirect should probably be redone.

Signed-off-by: Ian Allison <iana@pims.math.ca>