Finalizing changeover to farm domain
The final piece was the entityID in the shibboleth2.xml being used by
the shib image. I forgot to update this after updating the metadata so
was getting a message about our SP not being registered against the
staging IdP. Should be fixed now, but make sure you grab the latest
files from workspace when deploying.

This commit also changes the region for the deployed k8s1 instance to
Canada now that AWS has a Canadian region. Finally it tidies out the
output of the one-two-syzygy chart.

Signed-off-by: Ian Allison <iana@pims.math.ca>