Updating documentation and adding efs-mon
efs-mon is a basic ec2 instance ready to probe the EFS structure holding
our home directories. At the moment it is mostly manual config, the only
reason for including it here is so that resources can be tidied up
completely before attempting to tear down the eks cluster (otherwise
some items may be undeletable by terraform). I intend to extend it to
rely on remote state for information about the cluster (e.g. cluster
name and efs fs-id). When that is done more automation could usefully be added.

Signed-off-by: Ian Allison <iana@pims.math.ca>